Rename layers file to index.json
